Restructure Parallel/GatherScatter tests
