+++ tests/parallel/gather-scatter/EnforceGatherScatterEmptyReceivedMaster.cpp
@@ -1,0 +1,22 @@
+#ifndef PRECICE_NO_MPI
+
+#include "testing/Testing.hpp"
+
+#include <precice/SolverInterface.hpp>
+#include "helpers.hpp"
+
+BOOST_AUTO_TEST_SUITE(PreciceTests)
+BOOST_AUTO_TEST_SUITE(Parallel)
+BOOST_AUTO_TEST_SUITE(GatherScatter)
+BOOST_AUTO_TEST_CASE(EnforceGatherScatterEmptyReceivedMaster)
+{
+  PRECICE_TEST("ParallelSolver"_on(2_ranks), "SerialSolver"_on(1_rank));
+  // Provided master partition is not empty, but received master partitionis empty
+  runTestEnforceGatherScatter(std::vector<double>{0.0, 2.0, 0.0, 2.5}, context);
+}
+
+BOOST_AUTO_TEST_SUITE_END() // PreciceTests
+BOOST_AUTO_TEST_SUITE_END() // Parallel
+BOOST_AUTO_TEST_SUITE_END() // GatherScatter
+
+#endif // PRECICE_NO_MPI
